+++ bin/format.go
@@ -13,45 +13,9 @@ import (
 	"strings"
 )
 
-// Get the Queue type pretty format for output. [[k,  v]]
-// Deprecated: Use FormatKv instead.
-func FormatStr(d string, ss ...[][]string) string {
-	if d == "" {
-		// 4 个空格
-		d = "   "
-	}
-	bit := d[0:1]
-
-	// 或者最大长度
-	maxLen := 0
-	for _, sg := range ss {
-		for _, s := range sg {
-			kLen := len(s[0])
-			if kLen > maxLen {
-				maxLen = kLen
-			}
-		}
-	}
-
-	maxLen += len(d)
-
-	// 格式化
-	var contents string
-	for _, sg := range ss {
-		for _, s := range sg {
-			ss1 := s[0] + strings.Repeat(bit, maxLen-len(s[0])) + s[1] + "\n"
-			contents += ss1
-		}
-	}
-
-	return contents
-}
-
-/*
-The `k-v` data format to beautiful str.
-	FormatKv(kv map[string]interface{}, pref string)				 provide pref param form FormatKv.
-	FormatKv(kv map[string]interface{}, pref string, md string)	     provide pref and middle param form FormatK.
-*/
+// FormatKv The `k-v` data format to beautiful str.
+//FormatKv(kv map[string]interface{}, pref string)				 provide pref param form FormatKv.
+//FormatKv(kv map[string]interface{}, pref string, md string)	     provide pref and middle param form FormatK.
 func FormatKv(kv interface{}, params ...string) string {
 	var vf = reflect.ValueOf(kv)
 	if vf.Kind() != reflect.Map {
@@ -95,11 +59,9 @@ func FormatKv(kv interface{}, params ...string) string {
 	return s
 }
 
-/*
-The `k-v` data format to beautiful str.
-	FormatKvSort(kv map[string]interface{}, pref string)				 provide pref param form FormatKv.
-	FormatKvSort(kv map[string]interface{}, pref string, md string)	     provide pref and middle param form FormatK.
-*/
+// FormatKvSort The `k-v` data format to beautiful str.
+//FormatKvSort(kv map[string]interface{}, pref string)				 provide pref param form FormatKv.
+//FormatKvSort(kv map[string]interface{}, pref string, md string)	     provide pref and middle param form FormatK.
 func FormatKvSort(kv interface{}, params ...string) string {
 	var vf = reflect.ValueOf(kv)
 	if vf.Kind() != reflect.Map {
@@ -146,9 +108,8 @@ func FormatKvSort(kv interface{}, params ...string) string {
 	return s
 }
 
-// 格式化数组字符
-// 用于命令行输出
-// prefs 为 "" 时默认以数组索引开头；否则默给定的输出
+// FormatQue format the string array, using for cli output pretty.
+// where prefs is empty default use the array index
 func FormatQue(que interface{}, prefs ...string) string {
 	pref := ""  // 开头符号
 	dter := " " // 空格
@@ -181,10 +142,9 @@ func FormatQue(que interface{}, prefs ...string) string {
 	return s
 }
 
+// FormatTable Table format output by slice:
+// 	(table, bool) if is use the idx, table is 2 dimensional array.
 // Bug(FormatQue): chinese text cannot alignment
-//
-// Table format output by slice:
-// 	(data, bool) if is use the idx
 func FormatTable(table interface{}, args ...interface{}) string {
 	useIdxMk := true
 	if args != nil {