+++ doc/uymas-test-shell/base-script.sh
@@ -1,0 +1,17 @@
+# 2021年6月30日；uymas 加载其测试脚本
+# 使用 `$ --file,=f <script> 发起`
+
+# 测试版本
+test --name "Joshua Conero" --year=1992 --name "LiMing" --name "Emma" "And so on, yep is good."
+
+
+# 拼音
+pinyin 中华人民共和国
+
+
+# 版本
+version
+
+
+# 不存在的命令
+not-exit -xyz --name "不存在的命令    测试"